move gen_sql_stmt to simple_sqlite3_orm._utils module
